custom class separated from fragment
